requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.9.0
Django==4.2.6
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.109.2
filelock==3.14.0
flatbuffers==23.5.26
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
get-mnist==1.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.15.0
kiwisolver==1.4.5
langchain-community==0.0.12
langchain-core==0.1.10
langsmith==0.0.80
libclang==16.0.6
Mako==1.3.2
Markdown==3.5.1
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.2
oauthlib==3.2.2
openai==1.7.0
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
pandas==2.1.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.0.6
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
